add svgo-examples
