+++ examples/svgo-export/.gitignore
@@ -1,33 +1,13 @@
-# Logs
-logs
-*.log
-npm-debug.log*
+# build artifacts
+plugin.sketchplugin
 
-# Runtime data
-pids
-*.pid
-*.seed
-
-# Directory for instrumented libs generated by jscoverage/JSCover
-lib-cov
-
-# Coverage directory used by tools like istanbul
-coverage
-
-# Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)
-.grunt
-
-# node-waf configuration
-.lock-wscript
-
-# Compiled binary addons (http://nodejs.org/api/addons.html)
-build/Release
-
-# Dependency directory
+# npm
 node_modules
-
-# Optional npm cache directory
 .npm
+npm-debug.log
 
-# Optional REPL history
-.node_repl_history
+# mac
+.DS_Store
+
+# WebStorm
+.idea
