+++ examples/svgo-export/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "svgo_export",
+  "description": "Uses SVGO to compress exported SVG assets",
+  "author": "Ale Muñoz <ale@sketchapp.com>",
+  "version": "0.1.0",
+  "engines": {
+    "sketch": ">=3.0"
+  },
+  "skpm": {
+    "name": "SVGO Export",
+    "manifest": "src/manifest.json",
+    "main": "plugin.sketchplugin",
+    "assets": [
+      "assets/**/*"
+    ]
+  },
+  "scripts": {
+    "build": "skpm-build",
+    "watch": "skpm-build --watch",
+    "start": "skpm-build --watch --run",
+    "postinstall": "cd assets && npm i && cd .. && npm run build && skpm-link"
+  },
+  "devDependencies": {
+    "@skpm/builder": "^0.4.0"
+  },
+  "dependencies": {
+    "sketch-utils": "^0.1.2"
+  }
+}
