+++ examples/text-utilities/src/text-utilities.js
@@ -1,4 +1,4 @@
-// Text Utilities, by Johnnie Walker — Source code available at [GitHub](https://github.com/BohemianCoding/plugins.examples.text-utilities)
+// Text Utilities, by Johnnie Walker — Source code available at [GitHub](https://github.com/BohemianCoding/SketchAPI/tree/develop/examples/text-utilities)
 //
 // This plugin provides some debugging tools which annotate text layers to show where their baselines and bounding boxes are.
